move useSelectionState back again, re-reference
