+++ client/src/app/hooks/selection/useBulkSelection.ts
@@ -1,6 +1,6 @@
-import { IToolbarBulkSelectorProps } from "@app/components/ToolbarBulkSelector";
-import { useSelectionState } from "./useSelectionState";
 import { TdProps } from "@patternfly/react-table";
+import { IToolbarBulkSelectorProps } from "@app/components/ToolbarBulkSelector";
+import { useSelectionState } from "../useSelectionState/useSelectionState";
 
 export interface BulkSelectionArgs<ItemType> {
   /**
