+++ client/src/app/hooks/useSelectionState/useSelectionState.ts
